+++ Koco/src/main/java/icet/koco/alarm/service/AlarmService.java
@@ -1,5 +1,6 @@
 package icet.koco.alarm.service;
 
+import icet.koco.alarm.dto.AlarmListDto;
 import icet.koco.alarm.dto.AlarmListDto.AlarmDto;
 import icet.koco.alarm.dto.AlarmRequestDto;
 import icet.koco.alarm.emitter.EmitterRepository;
@@ -12,9 +13,11 @@ import icet.koco.user.entity.User;
 import icet.koco.user.repository.UserRepository;
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
@@ -22,10 +25,14 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 @Service
 @RequiredArgsConstructor
 public class AlarmService {
+
     private final AlarmRepository alarmRepository;
     private final UserRepository userRepository;
     private final PostRepository postRepository;
     private final EmitterRepository emitterRepository;
+
+    @Value("${BASE_URL}")
+    private String baseUrl;
 
     public void createAlarmInternal(AlarmRequestDto requestDto) {
         // 게시글, 알림 송수신자 찾기
@@ -85,5 +92,77 @@ public class AlarmService {
                 emitterRepository.deleteById(emitterId);
             }
         });
+    }
+
+    public SseEmitter subscribe(Long userId, String lastEventId) {
+        String emitterId = makeEmitterId(userId);
+        SseEmitter emitter = new SseEmitter(60L * 1000 * 60); // 1시간 유효
+
+        // emitter 등록
+        emitterRepository.save(emitterId, emitter);
+
+        // emitter 종료 처리 등록
+        emitter.onCompletion(() -> emitterRepository.deleteById(emitterId));
+        emitter.onTimeout(() -> emitterRepository.deleteById(emitterId));
+        emitter.onError((e) -> emitterRepository.deleteById(emitterId));
+
+        // 연결 성공 메시지 전송
+        try {
+            emitter.send(SseEmitter.event()
+                    .id(emitterId)
+                    .name("connect")
+                    .data("connected"));
+        } catch (IOException e) {
+            throw new RuntimeException("SSE 연결 실패", e);
+        }
+
+        // 서버 재시작으로 인한 이벤트 누락 복원 (Last-Event-ID 기준 캐시 재전송)
+        if (lastEventId != null) {
+            Map<String, Object> cachedEvents = emitterRepository.findAllEventCacheByUserId(String.valueOf(userId));
+            cachedEvents.entrySet().stream()
+                    .filter(entry -> entry.getKey().compareTo(lastEventId) > 0)
+                    .forEach(entry -> {
+                        try {
+                            emitter.send(SseEmitter.event()
+                                    .id(entry.getKey())
+                                    .name("alarm")
+                                    .data(entry.getValue()));
+                        } catch (IOException e) {
+                            emitterRepository.deleteById(emitterId);
+                        }
+                    });
+        }
+
+        // 읽지 않은 알림(DB에 isRead = false)을 재전송
+        List<Alarm> unreadAlarms = alarmRepository.findByReceiverIdAndIsReadFalse(userId);
+        unreadAlarms.forEach(alarm -> {
+            AlarmListDto.AlarmDto dto = AlarmListDto.AlarmDto.builder()
+                    .id(alarm.getId())
+                    .postId(alarm.getPost().getId())
+                    .postTitle(alarm.getPost().getTitle())
+                    .receiverId(alarm.getReceiver().getId())
+                    .senderId(alarm.getSender().getId())
+                    .senderNickname(alarm.getSender().getNickname())
+                    .alarmType(alarm.getAlarmType())
+                    .url(baseUrl + "/posts/" + alarm.getPost().getId())
+                    .createdAt(alarm.getCreatedAt())
+                    .build();
+
+            try {
+                emitter.send(SseEmitter.event()
+                        .id(emitterId)
+                        .name("alarm")
+                        .data(dto));
+                emitterRepository.saveEventCache(emitterId, dto); // 캐시 저장
+            } catch (IOException e) {
+                log.warn("❌ unread 알림 전송 실패: alarmId={}, error={}", alarm.getId(), e.getMessage());
+            }
+        });
+
+        return emitter;
+    }
+
+    String makeEmitterId(Long userId) {
+        return userId + "_" + System.currentTimeMillis();
     }
 }
